+++ controllers/user/delete-user.js
@@ -11,7 +11,18 @@ const deleteUser = async (req, res) => {
     try {
         const {id} = req.body;
 
-        const deletedUser = await prisma.prisma.user.delete({
+        const userForDelete = await prisma.prisma.user.findFirst({
+            where: {
+                id
+            }
+        })
+
+        // Если ни одного пользователя по заданному Id не удалось найти (вернулся null)
+        if (!userForDelete) {
+            return res.status(400).json({message: 'Пользователь не найден'});
+        }
+
+        await prisma.prisma.user.delete({
             where: {
                 id,
                 // админ может удалять только своего пользователя по id
@@ -19,7 +30,7 @@ const deleteUser = async (req, res) => {
             },
         });
 
-        if (deletedUser) {
+        if (userForDelete) {
             res.status(200).json({message: 'Пользователь удалён!'});
         }
